# Improve docker setup