--- a/backend/Dockerfile
+++ b/backend/Dockerfile
@@ -1,60 +1,36 @@
 # Use same version as defined in .python-version.
 FROM python:3.11.0-slim-bullseye
 
-# Update Ubuntu Software repository.
-RUN apt update -y \
-    && apt upgrade -y \
-    && apt install -y --no-install-recommends git build-essential procps curl file git nano sudo ncdu gcc postgresql libpq-dev \
-    && rm -rf /var/lib/apt/lists/*
-
-### Environment variables. ###
-ENV PYTHONDONTWRITEBYTECODE=1
-ENV PYTHONUNBUFFERED=1
-ENV PIPENV_VENV_IN_PROJECT=1
-# Docker should configure env vars in '.docker.env'.
-ENV PIPENV_DONT_LOAD_ENV=1
-
-# Set working directory.
-WORKDIR /app
-
-RUN echo 'alias la="ls -la"' >> ~/.bashrc
-
-# Upgrade pip.
-RUN python -m pip install --upgrade pip
-
-# Install pipenv.
-RUN python -m pip install pipenv
-
-# Copy dependency files into workdir (optimise docker cache layers).
-COPY Pipfile Pipfile.lock ./
-
-# Install virtual environment.
-RUN python -m pipenv install --dev --deploy
-
-# Copy rest of the project into image.
-COPY . .
-
-# Expose port for django and debugpy.
+ENV PYTHONUNBUFFERED 1
 EXPOSE 8000
 EXPOSE 5678
 
-# Update static and migration on each startup.
-ENTRYPOINT ["/app/entrypoint.sh"]
+# Alias
+RUN echo 'alias la="ls -la"' >> ~/.bashrc
 
-# Final command.
-# CMD ["python", "-m", "pipenv", "run", "python", "manage.py", "runserver", "0.0.0.0:8000"]
-CMD [ \
-  "python", \
-  "-m", \
-  "pipenv", \
-  "run", \
-  "gunicorn", \
-  "--bind=0.0.0.0:8000", \
-  "--forwarded-allow-ips=*", \
-  "--pythonpath=/app", \
-  "--workers=1", \
-  "--timeout=120", \
-  "--worker-class=gthread", \
-  "--threads=5", \
-  "root.wsgi:application" \
-]
+# Useful alias shortcuts for django
+RUN echo 'alias migrate="pipenv run python /app/manage.py migrate"' >> ~/.bashrc
+RUN echo 'alias makemigrations="pipenv run python /app/manage.py makemigrations"' >> ~/.bashrc
+RUN echo 'alias seed="pipenv run python /app/manage.py seed"' >> ~/.bashrc
+RUN echo 'alias collectstatic="pipenv run python /app/manage.py collectstatic --noinput"' >> ~/.bashrc
+
+# Make directories
+RUN mkdir /app
+WORKDIR /app
+
+# Docker should configure env vars in '.docker.env'.
+ENV PIPENV_DONT_LOAD_ENV=1
+
+# Copy dependencies
+COPY Pipfile Pipfile.lock ./
+
+# Install virtual environment.
+RUN python -m pip install pipenv
+RUN python -m pipenv install --dev --deploy
+
+# Copy remaining
+COPY . /app
+
+# Start
+ENTRYPOINT ["/app/entrypoint.sh"]
+CMD ["python", "-m", "pipenv", "run", "python", "manage.py", "runserver", "0.0.0.0:8000"]