--- a/backend/root/settings/base.py
+++ b/backend/root/settings/base.py
@@ -197,7 +197,6 @@ INSTALLED_APPS += [
 ################## LOGGING ##################
 
 # pylint: disable=wrong-import-position,wrong-import-order
-import logging.config  # noqa: E402
 
 from root.utils.json_formatter import JsonFormatter  # noqa: E402
 from root.custom_classes.request_context_filter import RequestContextFilter  # noqa: E402
@@ -303,7 +302,6 @@ LOGGING = {
         },
 }
 
-logging.config.dictConfig(LOGGING)
 
 # Quick fix for avoiding concurrency issues related to db access
 # Note: this might not be an ideal solution. See these links for information