- Improves build times for backend and frontend. On my system:
   - Backend:  115.9s -> 25.6s (4.5x faster)
   - Frontend: 262.1s -> 206.3s (1.3x faster) 
- Optimizes docker layering to prevent too many rebuilds of lower layers
- Add nice shortcuts in docker container for seeding/migrating etc
- Simplifies docker-compose and dockerfiles
- Add storybook to docker-compose  
   - No need to run manually anymore. Will of course not be needed in prod later on.
- Add nice little welcome splash
- Changes logging in backend dev environment
   - Stack traces were basically unreadable. Nice for prod, but not for dev. 
   
A lot of speed is gained by removing apt installs. These really go against what I think docker is great for, as it could break the project without changing anything. Instead we should use specific dependency versions and docker images if possible. Will have to look into this more when we enable cypress again.