@evgiz 2023-03-14
> Den funker dessverre ikke

Nå burde det fungere, var en bug med pipenv som hindrer den i å bruke riktig python path. 
Har også endret slik at storybook og frontend bruker samme image (raskere og tryggere mtp deps).

Kan du verifisere at det fungerer og at du får tilgang til både frontend og storybook? @emilte