@emilte 2023-03-14
frontend og storybook fungerer fortsatt dårlig for meg, men tror det er mac issue